# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Archive
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "0.1"
servers:
- url: http://localhost:8080
  description: Generated server url
tags:
- name: Archive resources
paths:
  /archive/byggr:
    post:
      tags:
      - Archive resources
      summary: Submit a ByggR archive request
      operationId: byggR
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ByggRArchiveRequest'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ArchiveResponse'
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Problem'
  /archive/byggr2:
    post:
      tags:
      - Archive resources
      summary: Submit a ByggR archive request with BASE64-encoded metadata
      operationId: byggR2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ByggRArchiveRequest2'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ArchiveResponse'
        "400":
          description: Bad request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Problem'
  /api-docs:
    get:
      tags:
      - API
      summary: OpenAPI-definition
      operationId: api-docs
      responses:
        "200":
          description: OK
      x-auth-type: None
      x-throttling-tier: Unlimited
components:
  schemas:
    Attachment:
      required:
      - extension
      - name
      type: object
      properties:
        name:
          type: string
        extension:
          pattern: ^\.(bmp|gif|tif|tiff|jpeg|jpg|png|htm|html|pdf|rtf|docx|txt|xlsx|odt|ods)$
          type: string
          example: .pdf
        file:
          type: string
          description: BASE64-encoded file contents
          example: YXBhCg==
      description: Bilaga
    ByggRArchiveRequest:
      required:
      - attachment
      - metadata
      type: object
      properties:
        attachment:
          $ref: '#/components/schemas/Attachment'
        metadata:
          type: string
          description: Metadata XML
      description: Request för att arkivera bygglovshandlingar
    ArchiveResponse:
      type: object
      properties:
        archiveId:
          type: string
          description: Archive ID
          nullable: true
        errorDetails:
          $ref: '#/components/schemas/ErrorDetails'
      description: Archive response
    ErrorDetails:
      type: object
      properties:
        errorCode:
          type: integer
          description: Formpipe error code
          format: int32
        errorMessage:
          type: string
          description: Formpipe error message
        serviceName:
          type: string
          description: Formpipe service name
          nullable: true
      description: Error details
      nullable: true
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: '#/components/schemas/StatusType'
        title:
          type: string
        detail:
          type: string
    StatusType:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        reasonPhrase:
          type: string
    ByggRArchiveRequest2:
      required:
      - attachment
      - metadata
      type: object
      properties:
        attachment:
          $ref: '#/components/schemas/Attachment'
        metadata:
          type: string
          description: Metadata XML
      description: "Request för att arkivera bygglovshandlingar, med BASE64-encode:ad\
        \ metadata"
  securitySchemes: {}